Separate Bootstapper into interface + implmentation.
