+++ xds/src/test/java/io/grpc/xds/BootstrapperImplTest.java
@@ -26,11 +26,15 @@ import io.grpc.TlsChannelCredentials;
 import io.grpc.internal.GrpcUtil;
 import io.grpc.internal.GrpcUtil.GrpcBuildVersion;
 import io.grpc.xds.Bootstrapper.BootstrapInfo;
+import io.grpc.xds.Bootstrapper.FileReader;
 import io.grpc.xds.Bootstrapper.ServerInfo;
 import io.grpc.xds.EnvoyProtoData.Locality;
 import io.grpc.xds.EnvoyProtoData.Node;
+import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -41,10 +45,32 @@ import org.junit.runners.JUnit4;
 @RunWith(JUnit4.class)
 public class BootstrapperTest {
 
+  private static final String BOOTSTRAP_FILE_PATH = "/home/grpc-java/xds/src/test/bootstrap.json";
   private static final String SERVER_URI = "trafficdirector.googleapis.com:443";
   @SuppressWarnings("deprecation") // https://github.com/grpc/grpc-java/issues/7467
   @Rule
   public final ExpectedException thrown = ExpectedException.none();
+
+  private final Bootstrapper bootstrapper = new Bootstrapper();
+  private String originalBootstrapPathFromEnvVar;
+  private String originalBootstrapPathFromSysProp;
+
+  @Before
+  public void setUp() {
+    saveEnvironment();
+    Bootstrapper.bootstrapPathFromEnvVar = BOOTSTRAP_FILE_PATH;
+  }
+
+  private void saveEnvironment() {
+    originalBootstrapPathFromEnvVar = Bootstrapper.bootstrapPathFromEnvVar;
+    originalBootstrapPathFromSysProp = Bootstrapper.bootstrapPathFromSysProp;
+  }
+
+  @After
+  public void restoreEnvironment() {
+    Bootstrapper.bootstrapPathFromEnvVar = originalBootstrapPathFromEnvVar;
+    Bootstrapper.bootstrapPathFromSysProp = originalBootstrapPathFromSysProp;
+  }
 
   @Test
   public void parseBootstrap_singleXdsServer() throws XdsInitializationException {
@@ -72,7 +98,8 @@ public class BootstrapperTest {
         + "  ]\n"
         + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     assertThat(info.getServers()).hasSize(1);
     ServerInfo serverInfo = Iterables.getOnlyElement(info.getServers());
     assertThat(serverInfo.getTarget()).isEqualTo(SERVER_URI);
@@ -123,7 +150,8 @@ public class BootstrapperTest {
         + "  ]\n"
         + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     assertThat(info.getServers()).hasSize(2);
     List<ServerInfo> serverInfoList = info.getServers();
     assertThat(serverInfoList.get(0).getTarget())
@@ -176,7 +204,8 @@ public class BootstrapperTest {
         + "  \"ignore\": \"something irrelevant\"\n"
         + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     assertThat(info.getServers()).hasSize(1);
     ServerInfo serverInfo = Iterables.getOnlyElement(info.getServers());
     assertThat(serverInfo.getTarget()).isEqualTo(SERVER_URI);
@@ -205,9 +234,10 @@ public class BootstrapperTest {
         + "  ]\n"
         + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     thrown.expect(XdsInitializationException.class);
     thrown.expectMessage("Invalid bootstrap: server " + SERVER_URI + " 'channel_creds' required");
-    Bootstrapper.parseConfig(rawData);
+    bootstrapper.bootstrap();
   }
 
   @Test
@@ -223,9 +253,10 @@ public class BootstrapperTest {
         + "  ]\n"
         + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     thrown.expect(XdsInitializationException.class);
     thrown.expectMessage("Server " + SERVER_URI + ": no supported channel credentials found");
-    Bootstrapper.parseConfig(rawData);
+    bootstrapper.bootstrap();
   }
 
   @Test
@@ -242,7 +273,8 @@ public class BootstrapperTest {
         + "  ]\n"
         + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     assertThat(info.getServers()).hasSize(1);
     ServerInfo serverInfo = Iterables.getOnlyElement(info.getServers());
     assertThat(serverInfo.getTarget()).isEqualTo(SERVER_URI);
@@ -268,9 +300,10 @@ public class BootstrapperTest {
         + "  }\n"
         + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     thrown.expect(XdsInitializationException.class);
     thrown.expectMessage("Invalid bootstrap: 'xds_servers' does not exist.");
-    Bootstrapper.parseConfig(rawData);
+    bootstrapper.bootstrap();
   }
 
   @Test
@@ -298,9 +331,9 @@ public class BootstrapperTest {
         + "  ]\n "
         + "}";
 
-    thrown.expect(XdsInitializationException.class);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     thrown.expectMessage("Invalid bootstrap: missing 'server_uri'");
-    Bootstrapper.parseConfig(rawData);
+    bootstrapper.bootstrap();
   }
 
   @Test
@@ -342,7 +375,8 @@ public class BootstrapperTest {
             + "  }\n"
             + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     assertThat(info.getServers()).isEmpty();
     assertThat(info.getNode()).isEqualTo(getNodeBuilder().build());
     Map<String, Bootstrapper.CertificateProviderInfo> certProviders = info.getCertProviders();
@@ -396,8 +430,9 @@ public class BootstrapperTest {
             + "  }\n"
             + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     try {
-      Bootstrapper.parseConfig(rawData);
+      bootstrapper.bootstrap();
       fail("exception expected");
     } catch (ClassCastException expected) {
       assertThat(expected).hasMessageThat().contains("value '234.0' for key 'plugin_name' in");
@@ -421,8 +456,9 @@ public class BootstrapperTest {
             + "  }\n"
             + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     try {
-      Bootstrapper.parseConfig(rawData);
+      bootstrapper.bootstrap();
       fail("exception expected");
     } catch (ClassCastException expected) {
       assertThat(expected).hasMessageThat().contains("value 'badValue' for key 'config' in");
@@ -445,8 +481,9 @@ public class BootstrapperTest {
             + "  }\n"
             + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     try {
-      Bootstrapper.parseConfig(rawData);
+      bootstrapper.bootstrap();
       fail("exception expected");
     } catch (XdsInitializationException expected) {
       assertThat(expected)
@@ -493,8 +530,9 @@ public class BootstrapperTest {
             + "  }\n"
             + "}";
 
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
     try {
-      Bootstrapper.parseConfig(rawData);
+      bootstrapper.bootstrap();
       fail("exception expected");
     } catch (XdsInitializationException expected) {
       assertThat(expected)
@@ -510,7 +548,8 @@ public class BootstrapperTest {
             + "  \"grpc_server_resource_name_id\": \"grpc/serverx\"\n"
             + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     assertThat(info.getGrpcServerResourceId()).isEqualTo("grpc/serverx");
   }
 
@@ -527,7 +566,9 @@ public class BootstrapperTest {
         + "    }\n"
         + "  ]\n"
         + "}";
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     ServerInfo serverInfo = Iterables.getOnlyElement(info.getServers());
     assertThat(serverInfo.getTarget()).isEqualTo(SERVER_URI);
     assertThat(serverInfo.getChannelCredentials()).isInstanceOf(InsecureChannelCredentials.class);
@@ -560,7 +601,8 @@ public class BootstrapperTest {
         + "  ]\n"
         + "}";
 
-    BootstrapInfo info = Bootstrapper.parseConfig(rawData);
+    bootstrapper.setFileReader(createFileReader(BOOTSTRAP_FILE_PATH, rawData));
+    BootstrapInfo info = bootstrapper.bootstrap();
     ServerInfo serverInfo = Iterables.getOnlyElement(info.getServers());
     assertThat(serverInfo.getTarget()).isEqualTo(SERVER_URI);
     assertThat(serverInfo.getChannelCredentials()).isInstanceOf(InsecureChannelCredentials.class);
@@ -569,6 +611,16 @@ public class BootstrapperTest {
     } else {
       assertThat(serverInfo.isUseProtocolV3()).isFalse();
     }
+  }
+
+  private static FileReader createFileReader(final String expectedPath, final String rawData) {
+    return new FileReader() {
+      @Override
+      public String readFile(String path) throws IOException {
+        assertThat(path).isEqualTo(expectedPath);
+        return rawData;
+      }
+    };
   }
 
   private static Node.Builder getNodeBuilder() {
